pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polily"
version = "0.1.0"
description = "Polymarket Decision Copilot — finds structure, shows direction lean, sizes risk"
requires-python = ">=3.11"
license = "MIT"
readme = "README.md"
keywords = ["polymarket", "prediction-market", "decision-support", "tui", "crypto"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Environment :: Console",
]

dependencies = [
    "httpx>=0.28,<1.0",
    "tenacity>=9.0,<10.0",
    "pydantic>=2.12,<3.0",
    "typer>=0.24,<1.0",
    "rich>=14.0,<15.0",
    "pyyaml>=6.0,<7.0",
    "ccxt>=4.0",
    "textual>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.3,<2.0",
    "pytest-cov>=5.0",
    "ruff>=0.15",
    "pyright>=1.1",
]

[tool.hatch.build.targets.wheel]
packages = ["scanner"]

[project.scripts]
polily = "scanner.cli:app"

[project.urls]
Homepage = "https://github.com/ShiyuCheng2018/polily"
Repository = "https://github.com/ShiyuCheng2018/polily"
Issues = "https://github.com/ShiyuCheng2018/polily/issues"

# ---------------------------------------------------------------------------
# Ruff (lint + format)
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["N806", "SIM102", "SIM108", "SIM117", "SIM118", "E501"]

[tool.ruff.format]
quote-style = "double"

# ---------------------------------------------------------------------------
# Pyright (type checking)
# ---------------------------------------------------------------------------
[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
